+++ docker-compose.mongo-replica.yml
@@ -1,0 +1,42 @@
+version: '3.8'
+
+services:
+
+  mongodb-primary:
+    image: mongo
+    command:
+      - --replSet
+      - rs0
+    volumes:
+      - $PWD/compose/mongo/primary/data/db:/data/db
+      - $PWD/compose/mongo/init:/docker-entrypoint-initdb.d:ro
+    expose:
+      - 27017
+    ports:
+      - 27018:27017
+    restart: always
+    networks:
+      default:
+        aliases:
+          - mxt.example
+
+  mongodb-secondary:
+    image: mongo
+    command: 
+      - --replSet
+      - rs0
+    volumes:
+      - $PWD/compose/mongo/secondary/data/db:/data/db
+    expose:
+      - 27017
+    restart: always
+
+  mongodb-arbiter:
+    image: mongo
+    command:
+      - --replSet
+      - rs0
+    expose:
+      - 27017
+    restart: always
+
